cleans up the generator - moves it into a package
