+++ packages/spec/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@supabase/generator",
+  "version": "0.0.0",
+  "main": "./index.tsx",
+  "types": "./index.tsx",
+  "license": "MIT",
+  "scripts": {
+    "cli:gen": "npm run cli:gen:commands && npm run cli:gen:config",
+    "cli:gen:commands": "ts-node ./index.ts gen --type command --input ../../spec/cli_v1_commands.yaml --output  ../../apps/reference/_cli/usage.mdx ",
+    "cli:gen:config": "ts-node ./index.ts gen --type config --input ../../spec/cli_v1_config.yaml --output  ../../apps/reference/_cli/config.mdx",
+    "gotrue:gen": "npm run gotrue:gen:config",
+    "gotrue:gen:config": "ts-node ./index.ts gen --type config --input ../../spec/gotrue_v1_config.yaml --output ../../apps/reference/_gotrue/config.mdx ",
+    "supabase-js:gen": "npm run supabase-js:gen:sdk",
+    "supabase-js:gen:sdk": "ts-node ./index.ts gen --type sdk --input ../../spec/supabase_js_sdk.yaml --output  ../../apps/reference/_supabase_js/usage.mdx",
+    "supabase-dart:gen": "npm run supabase-dart:gen:sdk",
+    "supabase-dart:gen:sdk": "ts-node ./index.ts gen --type sdk --input ../../spec/supabase_dart_sdk.yaml --output  ../../apps/reference/_supabase_dart/usage.mdx "
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "ejs": "^3.1.8",
+    "js-yaml": "^4.1.0",
+    "ts-node": "^10.9.1"
+  },
+  "prettier": {
+    "semi": false,
+    "singleQuote": true
+  }
+}
