+++ packages/spec/src/docs/cli.ts
@@ -1,4 +1,4 @@
-import template from '../templates/CliTemplate'
+import template from './templates/CliTemplate'
 import type { CliSpec } from '../types/CliSpec'
 
 const yaml = require('js-yaml')
@@ -7,19 +7,7 @@ const ejs = require('ejs')
 const Helpers = require('./Helpers')
 const { writeToDisk } = Helpers
 
-const main = (fileNames: string[], options: any) => {
-  try {
-    const outputDir = options.o || options.output || ''
-    fileNames.forEach((inputFileName) => {
-      gen(inputFileName, outputDir)
-    })
-    return
-  } catch (e) {
-    console.log(e)
-  }
-}
-
-async function gen(inputFileName: string, outputDir: string) {
+export default async function gen(inputFileName: string, outputDir: string) {
   const spec = yaml.load(fs.readFileSync(inputFileName, 'utf8'))
   // console.log('spec', spec)
 
@@ -33,10 +21,6 @@ async function gen(inputFileName: string, outputDir: string) {
       break
   }
 }
-
-// Run everything
-const argv = require('minimist')(process.argv.slice(2))
-main(argv['_'], argv)
 
 /**
  * Versioned Generator
