+++ packages/spec/src/docs/templates/CliTemplate.ts
