+++ packages/spec/src/docs/templates/ConfigTemplate.ts
