+++ packages/spec/src/types/CliSpec.ts
