+++ packages/spec/src/types/ConfigSpec.ts
