+++ packages/spec/src/types/Spec.ts
