+++ packages/spec/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "tsconfig/base.json",
+  "include": ["."],
+  "exclude": ["dist", "build", "node_modules"],
+  "compilerOptions": { "isolatedModules": false }
+}
