+++ spec/gotrue_v1_config.yaml
